Fix error if FFD is not installed.
